生成MapperXML

+++ izumiJava-demo/src/main/java/com/izumiJava/mappers/BaseMapper.java
@@ -8,22 +8,22 @@ public interface BaseMapper<T, P> {
     /**
      * insert:(插入)
      */
-    Integer insert(@Param("bean") T t);
+    Long insert(@Param("bean") T t);
 
     /**
      * insertOrUpdate:(插入或更新)
      */
-    Integer insertOrUpdate(@Param("bean") T t);
+    Long insertOrUpdate(@Param("bean") T t);
 
     /**
      * insertBatch:(批量插入)
      */
-    Integer insertBatch(@Param("list") List<T> list);
+    Long insertBatch(@Param("list") List<T> list);
 
     /**
      * insertOrUpdateBatch:(批量插入或更新)
      */
-    Integer insertOrUpdateBatch(@Param("list") List<T> list);
+    Long insertOrUpdateBatch(@Param("list") List<T> list);
 
     /**
      * selectList:(根据参数查询集合)
@@ -33,5 +33,5 @@ public interface BaseMapper<T, P> {
     /**
      * selectCount:(根据集合查询数量)
      */
-    Integer selectCount(@Param("query") P p);
+    Long selectCount(@Param("query") P p);
 }
